Add image and other attachment file in Ticket module

+++ app/Http/Controllers/TicketController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 use App\Http\Requests\StoreTicketRequest;
+use App\Http\Requests\UpdateTicketRequest;
 use Inertia\Inertia;
 use App\Models\Ticket;
 use App\Models\Customer;
@@ -95,10 +96,11 @@ class TicketController extends Controller
         $token = $this->getSavedToken();       
         $data = $request->validated();      
 
-        $request->validate([
-            'image' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-            'attach_file' => 'nullable|mimes:doc,docx,pdf,csv,xlx,xls|max:2048',
-        ]);
+        // $request->validate([
+        //     'image' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+        //     'attach_file' => 'nullable|mimes:doc,docx,pdf,csv,xlsx,xls|max:2048',
+        // ]);        
+
         $imageFileName = '';
         if($request->hasFile('image')) {
             $imageFileName = time() . '_'. $request->image->getClientOriginalName(); 
@@ -106,9 +108,9 @@ class TicketController extends Controller
         }
         
         $attachFileName = '';
-        if( $request->hasFile('attach_file') ) {
+        if($request->hasFile('attach_file')) {
             $attachFileName = time() . '_'. $request->attach_file->getClientOriginalName(); 
-            $request->image->move(public_path('uploads/others'), $attachFileName);
+            $request->attach_file->move(public_path('uploads/others'), $attachFileName);
         }        
 
         // return response(compact('fileName')); 
@@ -137,23 +139,53 @@ class TicketController extends Controller
         ]);
     } // edit
 
-    public function update(StoreTicketRequest $request, $id) 
+    public function update(UpdateTicketRequest $request, $id) 
     {
 		$input = $request->all();
         // return response(compact('input'));  
         $data = $request->validated(); 
-        $ticket = Ticket::findOrFail($id);        
-		
-        if($request->file('image')){            
-            if(file_exists(public_path('uploads/'.$ticket->image))){                
-                unlink(public_path('uploads/'.$ticket->image));
-            } 
+        $ticket = Ticket::findOrFail($id);  
+        
+        if($request->hasFile('image')) {
+            $request->validate([
+                'image' => 'nullable|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            ]); 
+        } 
+
+        if($request->hasFile('attach_file')) {
+            $request->validate([
+                'attach_file' => 'nullable|mimes:doc,docx,pdf,csv,xlsx,xls|max:2048',
+            ]); 
+        } 
+
+        if($request->file('image')){  
+            if($ticket->image) {
+                if(file_exists(public_path('uploads/'.$ticket->image))){                
+                    unlink(public_path('uploads/'.$ticket->image));
+                } 
+            }
+            
             $image = $request->file('image');
             $image_name = time() . '_'. $image->getClientOriginalName();
             $request->image->move(public_path('uploads'), $image_name);            
         } else {
             $image_name = $ticket->image;
         }
+
+        if($request->file('attach_file')){  
+            if($ticket->attach_file){
+                if(file_exists(public_path('uploads/others/'.$ticket->attach_file))){                
+                    unlink(public_path('uploads/others/'.$ticket->attach_file));
+                } 
+            }         
+            
+            $attach_file = $request->file('attach_file');
+            $attach_file_name = time() . '_'. $attach_file->getClientOriginalName();
+            $request->attach_file->move(public_path('uploads/others'), $attach_file_name);            
+        } else {
+            $attach_file_name = $ticket->attach_file;
+        }       
+       
         $ticket->update([
             'user_id' => $request->user_id,
             'ticket_source' => $request->ticket_source,             
@@ -163,7 +195,8 @@ class TicketController extends Controller
             'ticket_number'=> $request->ticket_number,
             'ticket_status' => $request->ticket_status,
             'updated_by_loggedin_user' => $request->updated_by_loggedin_user,
-            'image' => $image_name
+            'image' => $image_name,
+            'attach_file' => $attach_file_name,
         ]);
 		
         return redirect()->route('tickets')->with('status', 200); 
@@ -177,10 +210,46 @@ class TicketController extends Controller
                 unlink(public_path('uploads/'.$ticket->image));
             }                     
         } 
+
+        if($ticket->attach_file){
+            if(file_exists(public_path('uploads/others/'.$ticket->attach_file))){                
+                unlink(public_path('uploads/others/'.$ticket->attach_file));
+            }                     
+        } 
         // return response(compact('ticket'));
         $ticket->delete();
         return redirect()->route('tickets')->with('status', 204); 
     } // destroy
+
+    public function destroy_image($id)
+    {
+        $ticket = Ticket::findOrFail($id);
+        if($ticket->image){
+            if(file_exists(public_path('uploads/'.$ticket->image))){                
+                unlink(public_path('uploads/'.$ticket->image));
+            }
+            
+            $ticket->update([               
+                'image' => '',                
+            ]);
+        }        
+        return redirect()->route('tickets')->with('message', 'Image deleted successfully.'); 
+    } // destroy_image
+
+    public function destroy_attachFile($id)
+    {
+        $ticket = Ticket::findOrFail($id);
+        if($ticket->attach_file){
+            if(file_exists(public_path('uploads/others/'.$ticket->attach_file))){                
+                unlink(public_path('uploads/others/'.$ticket->attach_file));
+            }
+            
+            $ticket->update([               
+                'attach_file' => '',                
+            ]);
+        }        
+        return redirect()->route('tickets')->with('message', 'Attached file deleted successfully.'); 
+    } // destroy_attachFile
 
     public function store_remark(Request $request) { 
         $input = $request->all();
